More performance improvements and better resource management.

+++ Stockpile.h
@@ -33,7 +33,7 @@ public:
     int Paper;
     int Liquor;
     int Airship;
-    int Money;
+    long long Money;
 };
 
 #endif
